requirements.txt
delta-rest-client
pandas
stock-indicators
python-dotenv
requests